backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
google-generativeai
chromadb
langchain-text-splitters
redis
python-dotenv
pydantic-settings
pypdf
pdfplumber
python-docx
python-multipart
httpx
reportlab
pytest
pytest-asyncio
